requirements.txt
Django==5.1.1
djangorestframework==3.15.2
drf_yasg==1.21.10
python-dotenv==1.1.0
Requests==2.32.3
